CHATBOT/mangaba.ia/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mangaba"
version = "2.0.0"
description = "Framework de Agentes IA com Multi-Agent Orchestration, Protocolos A2A e MCP"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Mangaba AI Team", email = "contato@mangaba.ai"}
]
keywords = [
    "ai",
    "gemini",
    "google",
    "chatbot",
    "artificial-intelligence",
    "agent",
    "protocol"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"

dependencies = [
    "google-generativeai>=0.3.0",
    "openai>=1.6.0",
    "anthropic>=0.20.0",
    "huggingface-hub>=0.20.0",
    "python-dotenv>=0.19.0",
    "loguru>=0.6.0",
    "pydantic>=1.8.0",
    "requests>=2.25.0",
    "websockets>=10.0",
]

[project.optional-dependencies]
tools = [
    "duckduckgo-search>=3.9.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",
    "pytest-xdist>=3.3.1",
    "coverage>=7.2.0",
    "responses>=0.23.0",
    "factory-boy>=3.3.0",
    "faker>=19.0.0",
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.61.0",
    "freezegun>=1.2.2",
    "parameterized>=0.9.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest-html>=3.2.0",
    "pytest-json-report>=1.5.0",
    "pytest-github-actions-annotate-failures>=0.2.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",
    "pytest-xdist>=3.3.1",
    "coverage>=7.2.0",
]

[project.urls]
Homepage = "https://github.com/mangaba-ai/mangaba-ai"
"Bug Reports" = "https://github.com/mangaba-ai/mangaba-ai/issues"
Source = "https://github.com/mangaba-ai/mangaba-ai"
Documentation = "https://github.com/mangaba-ai/mangaba-ai#readme"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
# Configurações específicas para o gerenciador uv

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=html:htmlcov",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "--durations=10",
]
markers = [
    "unit: marca testes unitários",
    "integration: marca testes de integração",
    "performance: marca testes de performance",
    "slow: marca testes que demoram para executar",
    "api: marca testes que fazem chamadas de API",
    "network: marca testes que requerem conectividade",
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*",
    "setup.py",
    "conftest.py",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
multi_line_mode = 3
line_length = 120
skip = [".gitignore", ".dockerignore"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[tool.pylint]
max-line-length = 120
disable = [
    "C0111",  # missing-docstring
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
    "R0915",  # too-many-statements
]
